web：登录注册

+++ ofindex-frontend/src/components/account-cards/AccountCardRegister.vue
@@ -1,19 +1,35 @@
 <script setup>
 import {ref} from 'vue'
-import {Right} from "@element-plus/icons-vue";
+import {Back, Right} from "@element-plus/icons-vue";
 import {useRouter} from "vue-router";
 
 const router = useRouter();
-const AccountId = ref('');
+const NickName = ref('');
 const Password = ref('');
-const gotoRegister = function () {
-  router.push('/account/register');
+const ConfirmPassword = ref('');
+const goBack = function () {
+  router.back();
 }
-//发送登录请求
-const SendLoginReq = function () {
-  let token = "dsjakldjskaldjsal";
-  localStorage.setItem("Token", token);
-  router.push('/index');
+const sendRegisterReq = function () {
+  //判断两次输入的密码是否相同
+  if(Password.value !== ConfirmPassword.value) {
+      //弹窗
+      //清空输入框
+      Password.value = ConfirmPassword.value = '';
+  }
+  else{
+    //发送请求
+    let AccountId = '123123';
+    let token = 'ABCSDSJADKLJASJDKLSAJK';
+    localStorage.setItem('Token', token);
+    //跳转到成功页
+    router.push({
+      name: 'r-account-card-register-succeed',
+      params:{
+        AccountId:AccountId,
+      }
+    });
+  }
 }
 </script>
 
@@ -22,10 +38,10 @@ const SendLoginReq = function () {
     <div class="card-body center-layout-column">
       <el-row>
         <el-col :span="6" class="text-right">
-          账号：
+          昵称：
         </el-col>
         <el-col :span="18">
-          <el-input v-model="AccountId" placeholder="请输入账号" class="input-bar"></el-input>
+          <el-input v-model="NickName" placeholder="请输入昵称" class="input-bar"></el-input>
         </el-col>
       </el-row>
       <el-row>
@@ -36,16 +52,23 @@ const SendLoginReq = function () {
           <el-input v-model="Password" placeholder="请输入密码" class="input-bar"  type="password" show-password></el-input>
         </el-col>
       </el-row>
+      <el-row>
+        <el-col :span="6" class="text-right">
+          确认密码：
+        </el-col>
+        <el-col :span="18">
+          <el-input v-model="ConfirmPassword" placeholder="请再次输入密码" class="input-bar"  type="password" show-password></el-input>
+        </el-col>
+      </el-row>
     </div>
     <div class="card-footer">
       <el-row>
         <el-col :span="8" class="center-layout-row">
-          <el-button color="#80aaff" @click="gotoRegister">
-            注册
+          <el-button color="#80aaff" @click="goBack" :icon="Back">
           </el-button>
         </el-col>
         <el-col :span="8" :offset="8" class="center-layout-row">
-          <el-button color="#80aaff" @click="SendLoginReq" :icon="Right">
+          <el-button color="#80aaff" @click="sendRegisterReq" :icon="Right">
           </el-button>
         </el-col>
       </el-row>
